requirements.txt
asgiref==3.7.1
boto3==1.28.17
botocore==1.31.17
Django==4.2.1
django-cors-headers==4.0.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==21.2.0
jmespath==1.0.1
packaging==23.1
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.1
tzdata==2023.3
urllib3==1.26.16
whitenoise==6.5.0
